requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
prefect>=2.0
asgiref==3.11.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
blinker==1.9.0
cadwyn==6.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
constantly==23.10.4
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.5
cssselect==1.4.0
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.1
fastar==0.8.0
filelock==3.25.1
Flask==3.1.3
fonttools==4.62.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
llvmlite==0.46.0
lockfile==0.12.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
narwhals==2.17.0
numba==0.64.0
numpy==2.4.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==3.0.1
parsel==1.11.0
pathlib_abc==0.5.2
pathspec==1.0.4
pendulum==3.2.0
pillow==12.1.1
playwright==1.58.0
plotly==6.6.0
pluggy==1.6.0
Protego==0.6.0
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
cachetools==5.5.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
queuelib==1.9.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
retryhttp==1.4.0
rich==14.3.3
rich-argparse==1.7.2
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
Scrapy==2.14.1
seaborn==0.13.2
sentry-sdk==2.54.0
service-identity==24.2.0
setproctitle==1.3.7
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
soupsieve==2.8.3
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tldextract==5.3.1
tqdm==4.67.3
vaderSentiment==3.3.2
Twisted==25.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.41.0
w3lib==2.4.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wirerope==1.0.0
wrapt==2.1.2
xgboost==3.2.0
zipp==3.23.0
zope.interface==8.2
